+++ app/tests/tests.d.ts
@@ -1,0 +1,9 @@
+import type { SinonMock } from 'sinon'
+
+declare module 'mocha' {
+  export interface Context {
+    user: SinonMock
+    token: SinonMock
+    jwt: SinonMock
+  }
+}
